requirements.txt
django
restframeworks
